requirements.txt
mypy
black
isort
flake8
pytest
line_profiler

pyinstaller==5.2
numpy==1.23.1
numpy-quaternion==2022.4.2
bezier==2021.2.12
wxPython==4.1.1
Cython==0.29.32
PyOpenGL‑3.1.6‑cp39‑cp39‑win_amd64.whl
PyOpenGL_accelerate‑3.1.6‑cp39‑cp39‑win_amd64.whl
